refactor: subject cleanup
