+++ src/3-rxjs/play-subject/12-subject.ts
@@ -1,8 +1,4 @@
-// #region imports
-// tslint:disable:member-ordering
-// Namespace to get something you need
-import * as Rxjs from 'rxjs';
-// #endregion imports
+import { Subject } from 'rxjs';
 
 import { timerData$ } from '../core/helpers';
 
@@ -12,7 +8,7 @@ export function play(...args: any[]) {
    * Subject keep a list of subscribers and notify all of them (multi-cast)
    * each time they emit on a channel.
    */
-  const subject = new Rxjs.Subject();
+  const subject = new Subject();
 
   timerData$.subscribe(n => {
     subject.next(n);
